split file tests to file + cahnge folder names
